+++ src/subHeader/SubHeader.tsx
@@ -1,5 +1,6 @@
 import React from 'react'
 import Title from 'title'
+import cc from 'classcat'
 
 export interface SubHeaderProps {
   readonly className?: Classcat.Class
@@ -7,7 +8,7 @@ export interface SubHeaderProps {
 }
 
 const SubHeader = ({ className, children }: SubHeaderProps) => (
-  <Title className={className} headingLevel="2">
+  <Title className={cc(['kirk-subheader', className])} headingLevel="2">
     {children}
   </Title>
 )